hpsa: Remove (u8) from hpsa_show_volume_status to match upstream
Patchwork-id: 102602
O-Subject: [RHEL6.7 e-stor PATCH 48/50] hpsa: Remove (u8) from hpsa_show_volume_status to match upstream
Bugzilla: 1150660
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

Cleanup hpsa_show_volume_status so that it better matches upstream.
The (u8) was not in upstream commit.

+++ drivers/scsi/hpsa.c
@@ -1211,7 +1211,7 @@ static void hpsa_monitor_offline_device(struct ctlr_info *h,
 static void hpsa_show_volume_status(struct ctlr_info *h,
 	struct hpsa_scsi_dev_t *sd)
 {
-	if (sd->volume_offline == (u8) HPSA_VPD_LV_STATUS_UNSUPPORTED)
+	if (sd->volume_offline == HPSA_VPD_LV_STATUS_UNSUPPORTED)
 		dev_info(&h->pdev->dev,
 			"C%d:B%d:T%d:L%d Volume status is not available through vital product data pages.\n",
 			h->scsi_host->host_no,
